Split PartialOrd into another file
