+++ src/lazy.ts
@@ -1,11 +1,12 @@
 import { Eq, eqSymbol } from "./type-class/eq.js";
-import type { Ord, PartialOrd } from "./type-class/ord.js";
 
 import type { Applicative1 } from "./type-class/applicative.js";
 import type { Functor1 } from "./type-class/functor.js";
 import type { GetHktA1 } from "./hkt.js";
 import type { Monad1 } from "./type-class/monad.js";
+import type { Ord } from "./type-class/ord.js";
 import type { PartialEq } from "./type-class/partial-eq.js";
+import type { PartialOrd } from "./type-class/partial-ord.js";
 import type { Traversable1 } from "./type-class/traversable.js";
 
 const lazyNominal = Symbol("Lazy");
