+++ src/number.ts
@@ -1,7 +1,7 @@
 import { Option, none, some } from "./option.js";
 import { Ordering, equal, greater, less } from "./ordering.js";
 
-import { fromPartialCmp } from "./type-class/ord.js";
+import { fromPartialCmp } from "./type-class/partial-ord.js";
 
 export const partialCmp = (lhs: number, rhs: number): Option<Ordering> => {
     if (Number.isNaN(lhs) || Number.isNaN(rhs)) {
