+++ src/type-class/ord.ts
@@ -1,57 +1,13 @@
 import { Eq, eqSymbol } from "./eq.js";
-import { Option, flatMap, map, mapOr, some } from "../option.js";
-import { Ordering, and, equal, isEq } from "../ordering.js";
+import { Ordering, isEq } from "../ordering.js";
 
-import type { Contravariant } from "./variance.js";
-import type { Monoid } from "./monoid.js";
-import type { PartialEq } from "./partial-eq.js";
-
-declare const partialOrdNominal: unique symbol;
-export type PartialOrdHktKey = typeof partialOrdNominal;
+import type { PartialOrd } from "./partial-ord.js";
+import { some } from "../option.js";
 
 /**
- * All instances of `PartialOrd` must satisfy following conditions:
- * - Transitivity: If `f` is `PartialOrd`, for all `a`, `b` and `c`; `f(a, b) == f(b, c) == f(a, c)`.
- * - Duality: If `f` is `PartialOrd`, for all `a` and `b`; `f(a, b) == -f(b, a)`.
- */
-export interface PartialOrd<Lhs, Rhs = Lhs> extends PartialEq<Lhs, Rhs> {
-    readonly partialCmp: (lhs: Lhs, rhs: Rhs) => Option<Ordering>;
-}
-
-export const fromPartialCmp = <Lhs, Rhs>(
-    partialCmp: (lhs: Lhs, rhs: Rhs) => Option<Ordering>,
-): PartialOrd<Lhs, Rhs> => ({
-    partialCmp,
-    eq: (l, r) => mapOr(false)((order: Ordering) => isEq(order))(partialCmp(l, r)),
-});
-
-declare module "../hkt.js" {
-    interface HktDictA1<A1> {
-        [partialOrdNominal]: PartialOrd<A1>;
-    }
-}
-
-export const contravariant: Contravariant<PartialOrdHktKey> = {
-    contraMap: (f) => (ord) => fromPartialCmp((l, r) => ord.partialCmp(f(l), f(r))),
-};
-
-export const identity: PartialOrd<unknown> = fromPartialCmp(() => some(equal));
-
-export const monoid = <Lhs, Rhs>(): Monoid<PartialOrd<Lhs, Rhs>> => ({
-    combine: (x, y) => ({
-        partialCmp: (l, r) =>
-            flatMap((first: Ordering) =>
-                map((second: Ordering) => and(second)(first))(y.partialCmp(l, r)),
-            )(x.partialCmp(l, r)),
-        eq: (l, r) => x.eq(l, r) && y.eq(l, r),
-    }),
-    identity,
-});
-
-/**
- * All instances of `PartialOrd` must satisfy following conditions:
- * - Transitivity: If `f` is `PartialOrd`, for all `a`, `b` and `c`; `f(a, b) == f(b, c) == f(a, c)`.
- * - Duality: If `f` is `PartialOrd`, for all `a` and `b`; `f(a, b) == -f(b, a)`.
+ * All instances of `Ord` must satisfy following conditions:
+ * - Transitivity: If `f` is `Ord`, for all `a`, `b` and `c`; `f(a, b) == f(b, c) == f(a, c)`.
+ * - Duality: If `f` is `Ord`, for all `a` and `b`; `f(a, b) == -f(b, a)`.
  * - Strict: Ordering for all values is well-defined (so the return value is not an `Option`).
  */
 export interface Ord<Lhs, Rhs = Lhs> extends PartialOrd<Lhs, Rhs>, Eq<Lhs, Rhs> {
